requirements.txt
streamlit==1.28.1
google-generativeai==0.3.0
Pillow==10.1.0
